HttpSession & StompSession checks

+++ src/main/java/com/drc/server/websocket/WebSocketEventListener.java
@@ -1,78 +1,58 @@
 package com.drc.server.websocket;
 
-import com.drc.server.entity.ErrorMessage;
 import com.drc.server.entity.User;
-import com.drc.server.service.RoleService;
 import com.drc.server.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.messaging.SessionConnectEvent;
 import org.springframework.web.socket.messaging.SessionDisconnectEvent;
-
-import java.time.Instant;
 
 @RequiredArgsConstructor
 @Slf4j
 @Component
 public class WebSocketEventListener {
     private final WebSocketSessionRegistry sessionRegistry;
-    private final RoleService roleService;
     private final UserService userService;
-    private final SimpMessagingTemplate messagingTemplate;
-    private final String ERROR_DESTINATION = "/client/messages/";
 
     @EventListener
     public void handleWebSocketConnectListener(SessionConnectEvent event) {
         StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-        String sessionId = headerAccessor.getSessionId();
-        String username = headerAccessor.getFirstNativeHeader("username");
-        String roleStr = headerAccessor.getFirstNativeHeader("role");
-        log.debug("Session connected: sessionId: {} username: {} roleStr: {}", sessionId,username, roleStr );
-
-        if (username != null && roleStr != null) {
-            try
-            {
-                User user = new User(
-                        sessionId,
-                        username,
-                        0.0,
-                        roleService.getRoleByName(roleStr) // enum Role
-                );
-
-                String userValidationResult = userService.save(user);
-
-                if (userValidationResult.equals(UserService.VALIDATE_OK))
-                {
-                    sessionRegistry.register(sessionId, user);
-                    log.debug("Registered: sessionId={}, username={}, roleStr={}", sessionId, username, roleStr);
-                }
-                else {
-                    ErrorMessage error = new ErrorMessage("REGISTRATION_ERROR", userValidationResult, sessionId, username, Instant.now().toString());
-                    messagingTemplate.convertAndSend(ERROR_DESTINATION , error);
-                    log.debug("Failed to register user! Error: {}", error);
-                }
-
-            }
-            catch (Exception e) {
-                log.debug("Invalid user header values: {}", e.getMessage());
-            }
+        String sessionId =  (String) headerAccessor.getSessionAttributes().get(WebSocketHandshakeInterceptor.httpSessionIdParamName);
+        String username = userService.getUserByHttpSesssionid(sessionId).getName();
+        log.debug("SessionConnectEvent: sessionId: {} username: {}", sessionId, username);
+        log.debug("SessionConnectEvent attributes: {}", headerAccessor.getSessionAttributes());
+        String stompSessionId = headerAccessor.getSessionId();
+        log.info("STOMP CONNECT event, stompSessionId: {}", stompSessionId);
+        User user = userService.getUserByHttpSesssionid(sessionId);
+        user.setStompSessionId(stompSessionId);
+        userService.update(user);
+        if (user == null) {
+            log.debug("User not found for stompSessionId: {}", stompSessionId);
         } else {
-            log.debug("Missing user headers in WebSocket connect for session {}", sessionId);
+            log.debug("User found for stompSessionId: {}, user: {}", stompSessionId, user.getName());
         }
+        log.debug("Debug: ");
+        log.debug("user by stompSessionId: {} ", userService.getByStompSessionId(stompSessionId));
     }
 
     @EventListener
     public void handleWebSocketDisconnectListener(SessionDisconnectEvent event) {
-        StompHeaderAccessor headerAccessor = StompHeaderAccessor.wrap(event.getMessage());
-        User user = userService.getUserBySesssionid(headerAccessor.getSessionId());
-        log.debug("Client disconnected. Session ID: {}, User: {}", headerAccessor.getSessionId(), user);
-        sessionRegistry.unregister(event.getSessionId());
-        userService.delete(user);
-        log.debug("Unregistered and deleted user: {}" , user);
+        String stompSessionId = event.getSessionId();
+        log.debug("SessionDisconnectEvent: stompSessionId: {}", stompSessionId);
 
+        User user = userService.getByStompSessionId(stompSessionId);
+        if (user == null) {
+            log.debug("No user found for stompSessionId: {}", stompSessionId);
+            return;
+        }
+
+        log.debug("Disconnect user: {}", user);
+
+        sessionRegistry.unregister(user.getHttpSessionId());
+        userService.delete(user);
+        log.debug("Unregistered and deleted user: {}", user);
     }
 }
